+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- add context to logger in middlewares. This could be useful when logger support hooks.
+
 ## v4.0.1 - 07-09-2023
 
 ### Fixed